Lift IRAS/AKARI candidate-pair geometry into p9-core
