+++ crates/p9-core/src/coords/mod.rs
@@ -1,3 +1,4 @@
+pub mod candidate_pair;
 pub mod democratic_helio;
 pub mod observer;
 pub mod sky;
